Add package exports and fix file endings
