--- core/context.py
@@ -1,4 +1,0 @@
-from typing import Optional
-from services.rag_service import RAGService
-
-rag_service: Optional[RAGService] = None  # Will be set in main.py
