+++ core/scheduler.py
@@ -1,6 +1,6 @@
 import asyncio
 from datetime import datetime
-from core import context as ctx
+import core as ctx
 import logging
 
 logger = logging.getLogger(__name__)
