+++ services/__init__.py
@@ -5,12 +5,13 @@ Business logic services for AI Toolkit
 from .project_service import ProjectService
 from .llm_analyzer import LLMAnalyzer
 from .document_service import DocumentService
-from .dataset_service import DatasetService
+from .datasets.management_service import DatasetService
+from .rag_service import RAGService
 
 __all__ = [
     "ProjectService",
     "LLMAnalyzer", 
     "DocumentService",
     "DatasetService",
-    "RagService",
+    "RAGService",
 ]
