+++ services/datasets/__init__.py
@@ -1,0 +1,7 @@
+from .discovery_service import DatasetDiscoveryService
+from .management_service import DatasetService
+
+__all__ = [
+    "DatasetDiscoveryService",
+    "DatasetService",
+]
